@@ -1,19 +1,12 @@
 package provider
 
 import (
-	"strings"
-
-	"github.com/blang/semver"
 	"github.com/corymhall/pulumi-provider-pde/provider/pkg/provider/installers"
 	"github.com/corymhall/pulumi-provider-pde/provider/pkg/provider/local"
-	// "github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 
 	p "github.com/pulumi/pulumi-go-provider"
 	"github.com/pulumi/pulumi-go-provider/infer"
-	"github.com/pulumi/pulumi-go-provider/integration"
 	"github.com/pulumi/pulumi-go-provider/middleware/schema"
-	// "github.com/pulumi/pulumi/pkg/v3/resource/provider"
-	// pulumirpc "github.com/pulumi/pulumi/sdk/v3/proto/go"
 )
 
 // Version is initialized by the Go linker to contain the semver of this build.
@@ -24,11 +17,6 @@ const (
 )
 
 func NewProvider() p.Provider {
-
-	// provider.Main(Name, func(hc *provider.HostClient) (pulumirpc.ResourceProviderServer, error) {
-	//
-	// })
-
 	// We tell the provider what resources it needs to support.
 	// In this case, a single custom resource.
 	return infer.Provider(infer.Options{
@@ -42,25 +30,13 @@ func NewProvider() p.Provider {
 				},
 			},
 		},
-		// Components: []infer.InferredComponent{
-		// 	infer.Component[*local.Profile, local.ProfileArgs, *local.ProfileState](),
-		// },
-		// Functions: []infer.InferredFunction{
-		// 	infer.Function[*local.Profile, local.GetFileNameArgs, local.GetFileNameResult](),
-		// 	infer.Function[*local.Profile, pulumi.StringInput, local.GetFileNameResult](),
-		// },
 		Resources: []infer.InferredResource{
 			infer.Resource[*local.Link, local.LinkArgs, local.LinkState](),
 			infer.Resource[*local.File, local.FileArgs, local.FileState](),
 			infer.Resource[*installers.GitHubRelease, installers.GitHubReleaseInputs, installers.GitHubReleaseOutputs](),
 			infer.Resource[*installers.GitHubRepo, installers.GitHubRepoInputs, installers.GitHubRepoOutputs](),
+			infer.Resource[*installers.Shell, installers.ShellInputs, installers.ShellOutputs](),
 		},
 	})
 
-}
-func Schema(version string) (string, error) {
-	version = strings.TrimPrefix(version, "v")
-	s, err := integration.NewServer(Name, semver.MustParse(version), NewProvider()).
-		GetSchema(p.GetSchemaRequest{})
-	return s.Schema, err
 }
