@@ -1,4 +1,4 @@
-// Code generated by pulumi-gen-pde DO NOT EDIT.
+// Code generated by pulumi-language-go DO NOT EDIT.
 // *** WARNING: Do not edit by hand unless you're certain you know what you are doing! ***
 
 package installers
@@ -17,6 +17,7 @@ type GitHubRelease struct {
 
 	AssetName         pulumi.StringPtrOutput   `pulumi:"assetName"`
 	BinFolder         pulumi.StringPtrOutput   `pulumi:"binFolder"`
+	BinLocation       pulumi.StringPtrOutput   `pulumi:"binLocation"`
 	DownloadURL       pulumi.StringOutput      `pulumi:"downloadURL"`
 	Environment       pulumi.StringMapOutput   `pulumi:"environment"`
 	Executable        pulumi.StringPtrOutput   `pulumi:"executable"`
@@ -24,10 +25,10 @@ type GitHubRelease struct {
 	Interpreter       pulumi.StringArrayOutput `pulumi:"interpreter"`
 	Locations         pulumi.StringArrayOutput `pulumi:"locations"`
 	Org               pulumi.StringOutput      `pulumi:"org"`
+	ReleaseVersion    pulumi.StringPtrOutput   `pulumi:"releaseVersion"`
 	Repo              pulumi.StringOutput      `pulumi:"repo"`
 	UninstallCommands pulumi.StringArrayOutput `pulumi:"uninstallCommands"`
 	UpdateCommands    pulumi.StringArrayOutput `pulumi:"updateCommands"`
-	Version           pulumi.StringPtrOutput   `pulumi:"version"`
 }
 
 // NewGitHubRelease registers a new resource with the given unique name, arguments, and options.
@@ -78,9 +79,11 @@ func (GitHubReleaseState) ElementType() reflect.Type {
 type gitHubReleaseArgs struct {
 	AssetName         *string  `pulumi:"assetName"`
 	BinFolder         *string  `pulumi:"binFolder"`
+	BinLocation       *string  `pulumi:"binLocation"`
 	Executable        *string  `pulumi:"executable"`
 	InstallCommands   []string `pulumi:"installCommands"`
 	Org               string   `pulumi:"org"`
+	ReleaseVersion    *string  `pulumi:"releaseVersion"`
 	Repo              string   `pulumi:"repo"`
 	UninstallCommands []string `pulumi:"uninstallCommands"`
 	UpdateCommands    []string `pulumi:"updateCommands"`
@@ -90,9 +93,11 @@ type gitHubReleaseArgs struct {
 type GitHubReleaseArgs struct {
 	AssetName         pulumi.StringPtrInput
 	BinFolder         pulumi.StringPtrInput
+	BinLocation       pulumi.StringPtrInput
 	Executable        pulumi.StringPtrInput
 	InstallCommands   pulumi.StringArrayInput
 	Org               pulumi.StringInput
+	ReleaseVersion    pulumi.StringPtrInput
 	Repo              pulumi.StringInput
 	UninstallCommands pulumi.StringArrayInput
 	UpdateCommands    pulumi.StringArrayInput
@@ -193,6 +198,10 @@ func (o GitHubReleaseOutput) BinFolder() pulumi.StringPtrOutput {
 	return o.ApplyT(func(v *GitHubRelease) pulumi.StringPtrOutput { return v.BinFolder }).(pulumi.StringPtrOutput)
 }
 
+func (o GitHubReleaseOutput) BinLocation() pulumi.StringPtrOutput {
+	return o.ApplyT(func(v *GitHubRelease) pulumi.StringPtrOutput { return v.BinLocation }).(pulumi.StringPtrOutput)
+}
+
 func (o GitHubReleaseOutput) DownloadURL() pulumi.StringOutput {
 	return o.ApplyT(func(v *GitHubRelease) pulumi.StringOutput { return v.DownloadURL }).(pulumi.StringOutput)
 }
@@ -221,6 +230,10 @@ func (o GitHubReleaseOutput) Org() pulumi.StringOutput {
 	return o.ApplyT(func(v *GitHubRelease) pulumi.StringOutput { return v.Org }).(pulumi.StringOutput)
 }
 
+func (o GitHubReleaseOutput) ReleaseVersion() pulumi.StringPtrOutput {
+	return o.ApplyT(func(v *GitHubRelease) pulumi.StringPtrOutput { return v.ReleaseVersion }).(pulumi.StringPtrOutput)
+}
+
 func (o GitHubReleaseOutput) Repo() pulumi.StringOutput {
 	return o.ApplyT(func(v *GitHubRelease) pulumi.StringOutput { return v.Repo }).(pulumi.StringOutput)
 }
@@ -231,10 +244,6 @@ func (o GitHubReleaseOutput) UninstallCommands() pulumi.StringArrayOutput {
 
 func (o GitHubReleaseOutput) UpdateCommands() pulumi.StringArrayOutput {
 	return o.ApplyT(func(v *GitHubRelease) pulumi.StringArrayOutput { return v.UpdateCommands }).(pulumi.StringArrayOutput)
-}
-
-func (o GitHubReleaseOutput) Version() pulumi.StringPtrOutput {
-	return o.ApplyT(func(v *GitHubRelease) pulumi.StringPtrOutput { return v.Version }).(pulumi.StringPtrOutput)
 }
 
 type GitHubReleaseArrayOutput struct{ *pulumi.OutputState }
