@@ -1,4 +1,4 @@
-// Code generated by pulumi-gen-pde DO NOT EDIT.
+// Code generated by pulumi-language-go DO NOT EDIT.
 // *** WARNING: Do not edit by hand unless you're certain you know what you are doing! ***
 
 package installers
@@ -16,18 +16,16 @@ type GitHubRepo struct {
 	pulumi.CustomResourceState
 
 	AbsFolderName     pulumi.StringOutput      `pulumi:"absFolderName"`
-	AssetName         pulumi.StringPtrOutput   `pulumi:"assetName"`
 	Branch            pulumi.StringPtrOutput   `pulumi:"branch"`
 	Environment       pulumi.StringMapOutput   `pulumi:"environment"`
 	FolderName        pulumi.StringPtrOutput   `pulumi:"folderName"`
 	InstallCommands   pulumi.StringArrayOutput `pulumi:"installCommands"`
 	Interpreter       pulumi.StringArrayOutput `pulumi:"interpreter"`
-	Locations         pulumi.StringArrayOutput `pulumi:"locations"`
 	Org               pulumi.StringOutput      `pulumi:"org"`
 	Repo              pulumi.StringOutput      `pulumi:"repo"`
 	UninstallCommands pulumi.StringArrayOutput `pulumi:"uninstallCommands"`
 	UpdateCommands    pulumi.StringArrayOutput `pulumi:"updateCommands"`
-	Version           pulumi.StringPtrOutput   `pulumi:"version"`
+	Version           pulumi.StringOutput      `pulumi:"version"`
 }
 
 // NewGitHubRepo registers a new resource with the given unique name, arguments, and options.
@@ -187,10 +185,6 @@ func (o GitHubRepoOutput) AbsFolderName() pulumi.StringOutput {
 	return o.ApplyT(func(v *GitHubRepo) pulumi.StringOutput { return v.AbsFolderName }).(pulumi.StringOutput)
 }
 
-func (o GitHubRepoOutput) AssetName() pulumi.StringPtrOutput {
-	return o.ApplyT(func(v *GitHubRepo) pulumi.StringPtrOutput { return v.AssetName }).(pulumi.StringPtrOutput)
-}
-
 func (o GitHubRepoOutput) Branch() pulumi.StringPtrOutput {
 	return o.ApplyT(func(v *GitHubRepo) pulumi.StringPtrOutput { return v.Branch }).(pulumi.StringPtrOutput)
 }
@@ -211,10 +205,6 @@ func (o GitHubRepoOutput) Interpreter() pulumi.StringArrayOutput {
 	return o.ApplyT(func(v *GitHubRepo) pulumi.StringArrayOutput { return v.Interpreter }).(pulumi.StringArrayOutput)
 }
 
-func (o GitHubRepoOutput) Locations() pulumi.StringArrayOutput {
-	return o.ApplyT(func(v *GitHubRepo) pulumi.StringArrayOutput { return v.Locations }).(pulumi.StringArrayOutput)
-}
-
 func (o GitHubRepoOutput) Org() pulumi.StringOutput {
 	return o.ApplyT(func(v *GitHubRepo) pulumi.StringOutput { return v.Org }).(pulumi.StringOutput)
 }
@@ -231,8 +221,8 @@ func (o GitHubRepoOutput) UpdateCommands() pulumi.StringArrayOutput {
 	return o.ApplyT(func(v *GitHubRepo) pulumi.StringArrayOutput { return v.UpdateCommands }).(pulumi.StringArrayOutput)
 }
 
-func (o GitHubRepoOutput) Version() pulumi.StringPtrOutput {
-	return o.ApplyT(func(v *GitHubRepo) pulumi.StringPtrOutput { return v.Version }).(pulumi.StringPtrOutput)
+func (o GitHubRepoOutput) Version() pulumi.StringOutput {
+	return o.ApplyT(func(v *GitHubRepo) pulumi.StringOutput { return v.Version }).(pulumi.StringOutput)
 }
 
 type GitHubRepoArrayOutput struct{ *pulumi.OutputState }
