@@ -1,4 +1,4 @@
-// Code generated by pulumi-gen-pde DO NOT EDIT.
+// Code generated by pulumi-language-go DO NOT EDIT.
 // *** WARNING: Do not edit by hand unless you're certain you know what you are doing! ***
 
 package installers
@@ -17,14 +17,13 @@ type Shell struct {
 
 	BinLocation       pulumi.StringPtrOutput   `pulumi:"binLocation"`
 	DownloadURL       pulumi.StringOutput      `pulumi:"downloadURL"`
-	Environment       pulumi.StringMapOutput   `pulumi:"environment"`
 	Executable        pulumi.BoolPtrOutput     `pulumi:"executable"`
 	InstallCommands   pulumi.StringArrayOutput `pulumi:"installCommands"`
-	Interpreter       pulumi.StringArrayOutput `pulumi:"interpreter"`
-	Location          pulumi.StringOutput      `pulumi:"location"`
+	Location          pulumi.StringPtrOutput   `pulumi:"location"`
 	ProgramName       pulumi.StringOutput      `pulumi:"programName"`
 	UninstallCommands pulumi.StringArrayOutput `pulumi:"uninstallCommands"`
 	UpdateCommands    pulumi.StringArrayOutput `pulumi:"updateCommands"`
+	Version           pulumi.StringOutput      `pulumi:"version"`
 	VersionCommand    pulumi.StringPtrOutput   `pulumi:"versionCommand"`
 }
 
@@ -37,9 +36,6 @@ func NewShell(ctx *pulumi.Context,
 
 	if args.DownloadURL == nil {
 		return nil, errors.New("invalid value for required argument 'DownloadURL'")
-	}
-	if args.InstallCommands == nil {
-		return nil, errors.New("invalid value for required argument 'InstallCommands'")
 	}
 	if args.ProgramName == nil {
 		return nil, errors.New("invalid value for required argument 'ProgramName'")
@@ -77,24 +73,30 @@ func (ShellState) ElementType() reflect.Type {
 }
 
 type shellArgs struct {
-	BinLocation       *string  `pulumi:"binLocation"`
-	DownloadURL       string   `pulumi:"downloadURL"`
-	Executable        *bool    `pulumi:"executable"`
-	InstallCommands   []string `pulumi:"installCommands"`
-	ProgramName       string   `pulumi:"programName"`
-	UninstallCommands []string `pulumi:"uninstallCommands"`
-	UpdateCommands    []string `pulumi:"updateCommands"`
+	BinLocation       *string           `pulumi:"binLocation"`
+	DownloadURL       string            `pulumi:"downloadURL"`
+	Environment       map[string]string `pulumi:"environment"`
+	Executable        *bool             `pulumi:"executable"`
+	InstallCommands   []string          `pulumi:"installCommands"`
+	Interpreter       []string          `pulumi:"interpreter"`
+	ProgramName       string            `pulumi:"programName"`
+	UninstallCommands []string          `pulumi:"uninstallCommands"`
+	UpdateCommands    []string          `pulumi:"updateCommands"`
+	VersionCommand    *string           `pulumi:"versionCommand"`
 }
 
 // The set of arguments for constructing a Shell resource.
 type ShellArgs struct {
 	BinLocation       pulumi.StringPtrInput
 	DownloadURL       pulumi.StringInput
+	Environment       pulumi.StringMapInput
 	Executable        pulumi.BoolPtrInput
 	InstallCommands   pulumi.StringArrayInput
+	Interpreter       pulumi.StringArrayInput
 	ProgramName       pulumi.StringInput
 	UninstallCommands pulumi.StringArrayInput
 	UpdateCommands    pulumi.StringArrayInput
+	VersionCommand    pulumi.StringPtrInput
 }
 
 func (ShellArgs) ElementType() reflect.Type {
@@ -192,10 +194,6 @@ func (o ShellOutput) DownloadURL() pulumi.StringOutput {
 	return o.ApplyT(func(v *Shell) pulumi.StringOutput { return v.DownloadURL }).(pulumi.StringOutput)
 }
 
-func (o ShellOutput) Environment() pulumi.StringMapOutput {
-	return o.ApplyT(func(v *Shell) pulumi.StringMapOutput { return v.Environment }).(pulumi.StringMapOutput)
-}
-
 func (o ShellOutput) Executable() pulumi.BoolPtrOutput {
 	return o.ApplyT(func(v *Shell) pulumi.BoolPtrOutput { return v.Executable }).(pulumi.BoolPtrOutput)
 }
@@ -204,12 +202,8 @@ func (o ShellOutput) InstallCommands() pulumi.StringArrayOutput {
 	return o.ApplyT(func(v *Shell) pulumi.StringArrayOutput { return v.InstallCommands }).(pulumi.StringArrayOutput)
 }
 
-func (o ShellOutput) Interpreter() pulumi.StringArrayOutput {
-	return o.ApplyT(func(v *Shell) pulumi.StringArrayOutput { return v.Interpreter }).(pulumi.StringArrayOutput)
-}
-
-func (o ShellOutput) Location() pulumi.StringOutput {
-	return o.ApplyT(func(v *Shell) pulumi.StringOutput { return v.Location }).(pulumi.StringOutput)
+func (o ShellOutput) Location() pulumi.StringPtrOutput {
+	return o.ApplyT(func(v *Shell) pulumi.StringPtrOutput { return v.Location }).(pulumi.StringPtrOutput)
 }
 
 func (o ShellOutput) ProgramName() pulumi.StringOutput {
@@ -222,6 +216,10 @@ func (o ShellOutput) UninstallCommands() pulumi.StringArrayOutput {
 
 func (o ShellOutput) UpdateCommands() pulumi.StringArrayOutput {
 	return o.ApplyT(func(v *Shell) pulumi.StringArrayOutput { return v.UpdateCommands }).(pulumi.StringArrayOutput)
+}
+
+func (o ShellOutput) Version() pulumi.StringOutput {
+	return o.ApplyT(func(v *Shell) pulumi.StringOutput { return v.Version }).(pulumi.StringOutput)
 }
 
 func (o ShellOutput) VersionCommand() pulumi.StringPtrOutput {
