@@ -1,4 +1,4 @@
-// Code generated by pulumi-gen-pde DO NOT EDIT.
+// Code generated by pulumi-language-go DO NOT EDIT.
 // *** WARNING: Do not edit by hand unless you're certain you know what you are doing! ***
 
 package local
@@ -12,12 +12,16 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 )
 
+// A file projected into a pulumi resource
 type File struct {
 	pulumi.CustomResourceState
 
+	// The content of the file.
 	Content pulumi.StringArrayOutput `pulumi:"content"`
-	Force   pulumi.BoolOutput        `pulumi:"force"`
-	Path    pulumi.StringOutput      `pulumi:"path"`
+	// If an already existing file should be deleted if it exists.
+	Force pulumi.BoolOutput `pulumi:"force"`
+	// The path of the file.
+	Path pulumi.StringOutput `pulumi:"path"`
 }
 
 // NewFile registers a new resource with the given unique name, arguments, and options.
@@ -66,16 +70,22 @@ func (FileState) ElementType() reflect.Type {
 }
 
 type fileArgs struct {
+	// The content of the file.
 	Content []string `pulumi:"content"`
-	Force   *bool    `pulumi:"force"`
-	Path    string   `pulumi:"path"`
+	// If an already existing file should be deleted if it exists.
+	Force *bool `pulumi:"force"`
+	// The path of the file. This defaults to the name of the pulumi resource.
+	Path string `pulumi:"path"`
 }
 
 // The set of arguments for constructing a File resource.
 type FileArgs struct {
+	// The content of the file.
 	Content pulumi.StringArrayInput
-	Force   pulumi.BoolPtrInput
-	Path    pulumi.StringInput
+	// If an already existing file should be deleted if it exists.
+	Force pulumi.BoolPtrInput
+	// The path of the file. This defaults to the name of the pulumi resource.
+	Path pulumi.StringInput
 }
 
 func (FileArgs) ElementType() reflect.Type {
@@ -165,14 +175,17 @@ func (o FileOutput) ToFileOutputWithContext(ctx context.Context) FileOutput {
 	return o
 }
 
+// The content of the file.
 func (o FileOutput) Content() pulumi.StringArrayOutput {
 	return o.ApplyT(func(v *File) pulumi.StringArrayOutput { return v.Content }).(pulumi.StringArrayOutput)
 }
 
+// If an already existing file should be deleted if it exists.
 func (o FileOutput) Force() pulumi.BoolOutput {
 	return o.ApplyT(func(v *File) pulumi.BoolOutput { return v.Force }).(pulumi.BoolOutput)
 }
 
+// The path of the file.
 func (o FileOutput) Path() pulumi.StringOutput {
 	return o.ApplyT(func(v *File) pulumi.StringOutput { return v.Path }).(pulumi.StringOutput)
 }
